split handlers.py file
